requirements.txt
httpx==0.24.1
python-dotenv==1.0.0
pytz==2023.3
discord.py